move SQL runner as an internal class
